refactor: refactor heightmap page (#1759)
* refactor: refactor heightmap page


* refactor: remove unused import in printer/getters


---------
