+++ src/store/printer/types.ts
@@ -157,26 +157,30 @@ export interface PrinterStateFilamentSensors {
 }
 
 export interface PrinterStateBedMesh {
-    name: string
-    data: {
-        algo: string
-        max_x: number
-        max_y: number
-        mesh_x_pps: number
-        mesh_y_pps: number
+    profile_name: string
+    mesh_min: [number, number]
+    mesh_max: [number, number]
+    probed_matrix: number[][]
+    mesh_matrix: number[][]
+    profiles: {
+        [key: string]: PrinterStateBedMeshProfile
+    }
+}
+
+export interface PrinterStateBedMeshProfile {
+    points: number[][]
+    mesh_params: {
         min_x: number
+        max_x: number
         min_y: number
-        points: { [key: number]: number[] }
-        tension: number
-        version: number
+        max_y: number
         x_count: number
         y_count: number
+        mesh_x_pps: number
+        mesh_y_pps: number
+        algo: 'bicubic' | 'lagrange'
+        tension: number
     }
-    points: number[]
-    min: number
-    max: number
-    variance: number
-    is_active: boolean
 }
 
 export interface PrinterStateMacroParam {
